数据判断

+++ 08/huangyisan/message/user/views.py
@@ -37,9 +37,16 @@ def delete_user(request):
     #if len(models.get_messages()) == 1:
     #    return handle_error(request,'无法删除最后一个用户','func_error','user/list.html')
     #else:
-    username = request.GET.get('username','')
-    models.delete_user(username)
-    return HttpResponseRedirect('/user/list_user/')
+    print(models.get_exist_users_count())
+    if models.get_exist_users_count() == 1:
+        context = models.get_messages()
+        print(43,context)
+        return handle_error(request,'无法删除最后一个用户', 'func_error','user/list.html',context)
+        print('无法删除最后一个用户')
+    else:
+        username = request.GET.get('username','')
+        models.delete_user(username)
+        return HttpResponseRedirect('/user/list_user/')
 
 def edit_user(request):
     #if login_user == '':
@@ -61,55 +68,55 @@ def modify_user(request):
     tel = request.POST.get('tel')
     password = request.POST.get('password')
     print(username,age,tel,password)
-    models.modify_user(age,tel,password,username)
-    #if age < 0 or age > 150:
-    #    return handle_error(request,'这年龄很诡异，添加信息失败','data_error','user/edit.html')
-    #if len(tel) != 11:
-    #    return handle_error(request,'这手机号码很诡异，添加信息失败','data_error','user/edit.html')
-    #if len(password) < 6:
-    #    return handle_error(request,'密码设置过于简单，请重新设置','data_error','user/edit.html')
-    #else:
-    #    models.modify_user(name=name,age=age,tel=tel,password=password,users=models.get_messages())
-    return HttpResponseRedirect('/user/list_user/')
+    context = {'username':username, 'age':age, 'tel':tel, 'password':password}
+    #models.modify_user(age,tel,password,username)
+    if age < 0 or age > 150:
+        return handle_error(request,'这年龄很诡异，添加信息失败','data_error','user/edit.html',context)
+    if len(tel) != 11:
+        return handle_error(request,'这手机号码很诡异，添加信息失败','data_error','user/edit.html',context)
+    if len(password) < 6:
+        return handle_error(request,'密码设置过于简单，请重新设置','data_error','user/edit.html',context)
+    else:
+        #models.modify_user(name=username,age=age,tel=tel,password=password,users=models.get_messages())
+        models.modify_user(age,tel,password,username)
+        return HttpResponseRedirect('/user/list_user/')
 
 def add_func(request):
-    username = request.POST.get('username','')
+    username = request.POST.get('username','Null')
     age = request.POST.get('age','-1')
     age = int(age)
     tel = request.POST.get('tel','')
     password = request.POST.get('password','')
-    print(username,age,tel,password)
-    models.add_user(username,age,tel,password)
-    #if name in models.get_messages():
-    #    return handle_error(request,'此用户已经存在，添加信息失败','data_error','user/list.html')
-    #    '''
-    #    info_dict=models.get_messages()
-    #    #当添加一个已经存在的用户的时候，添加一个error的key和对应的'此用户已经存在'的值给原字典，一起传递给list.html进行渲染。
-    #    #因为单纯的渲染error不够，所以和数据需要一起传递。
-    #    info_dict['data_error'] = '此用户已经存在，添加信息失败'
-    #    context={'messages':info_dict}
-    #    #因为add行为在list.html。所以这里直接request了list.html。如果独立出一个界面再重定向回list.html会有问题。当出现error的时候，会因为重定向导致渲染丢失。
-    #    #总之这里我是强行要把add和list写同个界面了。。应该有更好的处理方法,来解决重定向后渲染丢失问题。
-    #    return render(request,'user/list.html',context)
-    #    '''
-    #if age < 0 or age > 150:
-    #    return handle_error(request,'这年龄很诡异，添加信息失败','data_error','user/list.html')
-    #if len(tel) != 11:
-    #    return handle_error(request,'这手机号码很诡异，添加信息失败','data_error','user/list.html')
-    #if len(password) < 6:
-    #    return handle_error(request,'密码设置过于简单，请重新设置','data_error','user/list.html')
-    #else:
-    #    models.modify_user(name=name,age=age,tel=tel,password=password,users=models.get_messages())
-    return HttpResponseRedirect('/user/list_user/')
+    #组织context信息，传递给handle_error函数，用于错误的时候用用户信息+错误信息对页面渲染
+    context = {'username':username, 'age':age, 'tel':tel, 'password':password}
+    if models.judege_user_exist(username):
+        print('已经存在的用户')
+        return handle_error(request,'已经存在的用户,添加信息失败','data_error','user/add.html',context)
+    if age < 0 or age > 150:
+        print('年龄异常')
+        return handle_error(request,'年龄异常,添加信息失败','data_error','user/add.html',context)
+    if len(tel) != 11:
+        print('手机号码异常')
+        return handle_error(request,'手机号码异常,添加信息失败','data_error','user/add.html',context)
+    if len(password) < 6:
+        print('密码过短')
+        return handle_error(request,'密码过短,添加信息失败','data_error','user/add.html',context)
+    else:
+        models.add_user(username,age,tel,password)
+        return HttpResponseRedirect('/user/list_user/')
 
 def add_user(request):
     return render(request,'user/add.html')
 
-def handle_error(request,error_message,kind_error,url):
-    info_dict=models.get_messages()
-    info_dict[kind_error] = error_message
-    context={'messages':info_dict}
-    print(context)
+def handle_error(request,error_message,kind_error,url,context):
+    if type(context) is dict:
+        print('这是个字典')
+        context[kind_error] = error_message
+        context={'messages':context}
+    elif type(context) is tuple:
+        context[0][kind_error] = error_message
+        context={'messages':context}
+        print(context)
     return render(request,url,context)
 
 def unauth_error(request):
